+++ sdk-java/src/main/java/ly/count/sdk/java/User.java
@@ -16,7 +16,7 @@ public abstract class User {
     public enum Gender {
         FEMALE("F"), MALE("M");
 
-        private String value;
+        private final String value;
 
         Gender(String value) {
             this.value = value;