+++ sdk-java/src/main/java/ly/count/sdk/java/internal/UniqueTimeGenerator.java
@@ -8,7 +8,7 @@ import java.util.List;
  * Unique timer, keeps last 10 returned values in memory.
  */
 class UniqueTimeGenerator implements Device.TimeGenerator {
-    List<Long> lastTsMs = new ArrayList<>(10);
+    final List<Long> lastTsMs = new ArrayList<>(10);
     long addition = 0;
 
     long currentTimeMillis() {